[nextest] move TestBinary::from_messages into test_list.rs
